# openapi/sigac-v1.yaml
openapi: 3.1.0
info:
  title: SIGAC API
  version: 0.3.23
servers:
  - url: /api/v1
paths:
...
  /agenda-imports:
    post:
      operationId: importAgenda
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Agenda import processed successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgendaImportResponse' }
        '400': { $ref: '#/components/responses/HttpValidationError' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/AgendaPermissionDenied' }
        '409': { $ref: '#/components/responses/IdempotencyKeyReused' }
        '413': { $ref: '#/components/responses/AgendaUploadTooLarge' }
        '415': { $ref: '#/components/responses/AgendaArtifactUnsupported' }
        '422': { $ref: '#/components/responses/AgendaLayoutRejected' }
        '500': { $ref: '#/components/responses/AgendaImportFailed' }
        '504': { $ref: '#/components/responses/AgendaImportTimeout' }
    get:
      operationId: listAgendaImports
      parameters:
        - in: query
          name: agendaDate
          required: false
          description: Filter by agenda date (YYYY-MM-DD).
          schema: { type: string, format: date }
        - in: query
          name: cursor
          required: false
          description: Opaque cursor returned by the preceding page.
          schema: { type: string }
        - in: query
          name: limit
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Cursor-paginated agenda import history
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgendaImportHistoryPage' }
        '400': { $ref: '#/components/responses/HttpValidationError' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/AgendaPermissionDenied' }
  /agenda-imports/{importacionId}:
    get:
      operationId: getAgendaImportDetail
      parameters:
        - in: path
          name: importacionId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Full detail of an agenda importación
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgendaImportDetail' }
        '400': { $ref: '#/components/responses/HttpValidationError' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/AgendaPermissionDenied' }
        '404': { $ref: '#/components/responses/AgendaImportNotFound' }
  /agenda-imports/{importacionId}/incidents:
    get:
      operationId: getAgendaImportIncidents
      parameters:
        - in: path
          name: importacionId
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: cursor
          required: false
          description: Opaque cursor returned by the preceding page.
          schema: { type: string }
        - in: query
          name: limit
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Cursor-paginated incidents for an agenda importación
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgendaImportIncidentsPage' }
        '400': { $ref: '#/components/responses/HttpValidationError' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/AgendaPermissionDenied' }
  /agendas/{date}:
    get:
      operationId: getAgendaDaySummary
      parameters:
        - in: path
          name: date
          required: true
          schema: { type: string, format: date }
      responses:
        '200':
          description: Day-level agenda summary read model
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgendaDayReadModel' }
        '400': { $ref: '#/components/responses/HttpValidationError' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/AgendaPermissionDenied' }
        '404': { $ref: '#/components/responses/AgendaNotFound' }
  /agendas/{date}/preparation-items:
    get:
      operationId: getAgendaPreparationList
      parameters:
        - in: path
          name: date
          required: true
          schema: { type: string, format: date }
        - in: query
          name: order
          required: false
          schema:
            type: string
            enum: [APPOINTMENT_TIME_ASC, PATIENT_NAME_ASC]
        - in: query
          name: cursor
          required: false
          description: Opaque cursor returned by the preceding page.
          schema: { type: string }
        - in: query
          name: limit
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Cursor-paginated agenda preparation items
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgendaPreparationPage' }
        '400': { $ref: '#/components/responses/HttpValidationError' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/AgendaPermissionDenied' }
  /agendas/{date}/preparation-items/print:
    get:
      operationId: printAgendaPreparationList
      parameters:
        - in: path
          name: date
          required: true
          schema: { type: string, format: date }
        - in: query
          name: order
          required: false
          schema:
            type: string
            enum: [APPOINTMENT_TIME_ASC, PATIENT_NAME_ASC]
      responses:
        '200':
          description: Full (non-paginated) preparation list for print
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgendaPreparationPrintResponse' }
        '400': { $ref: '#/components/responses/HttpValidationError' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/AgendaPermissionDenied' }
  /agendas/{date}/preparation-report:
    post:
      operationId: generatePreparationReport
      summary: Genera y descarga el PDF de lista de preparación para Archivo Clínico
      description: |
        Genera un PDF operativo agrupado por servicio y médico, listo para imprimir.
        El PDF se genera en memoria y se descarga directamente (sin persistencia).
        Requiere AGENDA_VIEW + AGENDA_PRINT.
        La fecha y el tenant se obtienen del contexto de autorización; nunca del body.
        Ref: preparation-reports spec REQ-PR-002, ADR-0030, ADR-0031.
      parameters:
        - in: path
          name: date
          required: true
          schema: { type: string, format: date, example: "2026-08-26" }
        - in: query
          name: order
          required: false
          description: Ordenamiento de citas. Default APPOINTMENT_TIME_ASC.
          schema:
            type: string
            enum: [APPOINTMENT_TIME_ASC, PATIENT_NAME_ASC]
        - in: query
          name: services
          required: false
          description: |
            Lista de códigos de servicio separados por coma, ej. "CL05,CL06".
            Omitir o dejar vacío para incluir todos los servicios.
          schema: { type: string }
      responses:
        '200':
          description: PDF generado correctamente, descarga directa
          headers:
            Content-Disposition:
              description: attachment; filename="lista-preparacion-{date}.pdf"
              schema: { type: string }
          content:
            application/pdf:
              schema: { type: string, format: binary }
        '400': { $ref: '#/components/responses/HttpValidationError' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/AgendaPermissionDenied' }
        '422': { $ref: '#/components/responses/NoActiveAppointments' }
        '500': { $ref: '#/components/responses/AgendaImportFailed' }
components:
...
  schemas:
...
    HttpFieldError:
      type: object
      additionalProperties: false
      required: [field, code]
      properties:
        field: { type: string }
        code: { type: string, enum: [REQUIRED, INVALID_FORMAT, INVALID_TYPE, OUT_OF_RANGE] }
    ProblemDetails:
      type: object
      additionalProperties: false
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        code:
          type: string
          enum: [HTTP_VALIDATION_ERROR, AUTHENTICATION_REQUIRED, PERMISSION_DENIED, INSUFFICIENT_ENABLING_SOURCE, EXPEDIENTE_NOT_FOUND, OPTIMISTIC_LOCK_CONFLICT, REQUEST_INVALID_TRANSITION, AGENDA_IMPORT_NOT_FOUND, AGENDA_NOT_FOUND, IDEMPOTENCY_KEY_REUSED, AGENDA_UPLOAD_TOO_LARGE, AGENDA_ARTIFACT_UNSUPPORTED, AGENDA_LAYOUT_REJECTED, AGENDA_IMPORT_FAILED, AGENDA_IMPORT_TIMEOUT]
        detail: { type: string }
        errors:
          type: array
          items: { $ref: '#/components/schemas/HttpFieldError' }
    AgendaImportMetrics:
      type: object
      additionalProperties: false
      required: [receivedRecords, processed, added, updated, unchanged, restored, pendingReview, rejected, duplicateFolio, withdrawnFromAgenda, incidents, errors]
      properties:
        receivedRecords: { type: integer, minimum: 0 }
        processed: { type: integer, minimum: 0 }
        added: { type: integer, minimum: 0 }
        updated: { type: integer, minimum: 0 }
        unchanged: { type: integer, minimum: 0 }
        restored: { type: integer, minimum: 0 }
        pendingReview: { type: integer, minimum: 0 }
        rejected: { type: integer, minimum: 0 }
        duplicateFolio: { type: integer, minimum: 0 }
        withdrawnFromAgenda: { type: integer, minimum: 0 }
        incidents: { type: integer, minimum: 0 }
        errors: { type: integer, minimum: 0 }
    AgendaImportSummary:
      type: object
      additionalProperties: false
      required: [importacionId, agendaDate, importedAt, outcome, metrics]
      properties:
        importacionId: { type: string, format: uuid }
        agendaDate: { type: string, format: date }
        importedAt: { type: string, format: date-time }
        outcome:
          type: string
          enum: [IMPORTED, ALREADY_IMPORTED, RECONCILED]
        metrics: { $ref: '#/components/schemas/AgendaImportMetrics' }
    AgendaImportResponse:
      type: object
      additionalProperties: false
      required: [importacionId, agendaDate, importedAt, outcome, metrics]
      properties:
        importacionId: { type: string, format: uuid }
        agendaDate: { type: string, format: date }
        importedAt: { type: string, format: date-time }
        outcome:
          type: string
          enum: [IMPORTED, ALREADY_IMPORTED, RECONCILED]
        metrics: { $ref: '#/components/schemas/AgendaImportMetrics' }
    AgendaImportHistoryPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/AgendaImportSummary' }
        nextCursor: { type: [string, 'null'] }
    RegistroImportadoResult:
      type: object
      additionalProperties: false
      required: [registroId, sourcePosition, folio, processingResult, incidentCodes]
      properties:
        registroId: { type: string, format: uuid }
        sourcePosition: { type: integer, minimum: 1 }
        folio: { type: [string, 'null'] }
        processingResult:
          type: string
          enum: [ADDED, UPDATED, UNCHANGED, RESTORED, PENDING_REVIEW, REJECTED, DUPLICATE_FOLIO]
        incidentCodes:
          type: array
          items:
            type: string
            enum: [PHYSICIAN_NOT_RESOLVED, PHYSICIAN_AMBIGUOUS, SERVICE_NOT_RESOLVED, EXPEDIENT_NOT_RESOLVED, REQUIRED_DATA_MISSING, ROW_INCONSISTENT, DUPLICATE_FOLIO_IN_SNAPSHOT]
    AgendaImportDetail:
      type: object
      additionalProperties: false
      required: [summary, registros]
      properties:
        summary: { $ref: '#/components/schemas/AgendaImportSummary' }
        registros:
          type: array
          items: { $ref: '#/components/schemas/RegistroImportadoResult' }
    AgendaImportIncidentSummary:
      type: object
      additionalProperties: false
      required: [incidenciaId, registroId, sourcePosition, type]
      properties:
        incidenciaId: { type: string, format: uuid }
        registroId: { type: string, format: uuid }
        sourcePosition: { type: integer, minimum: 1 }
        type:
          type: string
          enum: [PHYSICIAN_NOT_RESOLVED, PHYSICIAN_AMBIGUOUS, SERVICE_NOT_RESOLVED, EXPEDIENT_NOT_RESOLVED, REQUIRED_DATA_MISSING, ROW_INCONSISTENT, DUPLICATE_FOLIO_IN_SNAPSHOT]
    AgendaImportIncidentsPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/AgendaImportIncidentSummary' }
        nextCursor: { type: [string, 'null'] }
    AgendaDayReadModel:
      type: object
      additionalProperties: false
      required: [agendaDate, latestImportacionId, latestImportedAt, latestOutcome, activeAppointments, physicians, services, incidentCount]
      properties:
        agendaDate: { type: string, format: date }
        latestImportacionId: { type: string, format: uuid }
        latestImportedAt: { type: string, format: date-time }
        latestOutcome:
          type: string
          enum: [IMPORTED, ALREADY_IMPORTED, RECONCILED]
        activeAppointments: { type: integer, minimum: 0 }
        physicians: { type: integer, minimum: 0 }
        services: { type: integer, minimum: 0 }
        incidentCount: { type: integer, minimum: 0 }
...
    AgendaPreparationPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/AgendaPreparationItem' }
        nextCursor: { type: [string, 'null'] }
    AgendaPreparationPrintResponse:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/AgendaPreparationItem' }
  responses:
    HttpValidationError:
      description: Structurally invalid or malformed request
      content:
        application/problem+json:
          schema:
            allOf:
              - { $ref: '#/components/schemas/ProblemDetails' }
              - type: object
                required: [detail]
                properties:
                  status: { const: 400 }
                  code: { const: HTTP_VALIDATION_ERROR }
    AuthenticationRequired:
      description: Authentication is required
      content:
        application/problem+json:
          schema:
            allOf:
              - { $ref: '#/components/schemas/ProblemDetails' }
              - type: object
                properties:
                  status: { const: 401 }
                  code: { const: AUTHENTICATION_REQUIRED }
...
    AgendaPermissionDenied:
      description: Authenticated actor lacks AGENDA_IMPORT or AGENDA_VIEW
      content:
        application/problem+json:
          schema:
            allOf:
              - { $ref: '#/components/schemas/ProblemDetails' }
              - type: object
                properties:
                  status: { const: 403 }
                  code: { const: PERMISSION_DENIED }
    AgendaImportNotFound:
      description: Agenda importación not found in active tenant
      content:
        application/problem+json:
          schema:
            allOf:
              - { $ref: '#/components/schemas/ProblemDetails' }
              - type: object
                properties:
                  status: { const: 404 }
                  code: { const: AGENDA_IMPORT_NOT_FOUND }
    AgendaNotFound:
      description: Agenda not found for this date in active tenant
      content:
        application/problem+json:
          schema:
            allOf:
              - { $ref: '#/components/schemas/ProblemDetails' }
              - type: object
                properties:
                  status: { const: 404 }
                  code: { const: AGENDA_NOT_FOUND }
    IdempotencyKeyReused:
      description: Idempotency-Key already used with a different artifact
      content:
        application/problem+json:
          schema:
            allOf:
              - { $ref: '#/components/schemas/ProblemDetails' }
              - type: object
                properties:
                  status: { const: 409 }
                  code: { const: IDEMPOTENCY_KEY_REUSED }
    AgendaUploadTooLarge:
      description: Uploaded file exceeds the configured limit
      content:
        application/problem+json:
          schema:
            allOf:
              - { $ref: '#/components/schemas/ProblemDetails' }
              - type: object
                properties:
                  status: { const: 413 }
                  code: { const: AGENDA_UPLOAD_TOO_LARGE }
    AgendaArtifactUnsupported:
      description: Artifact format or extension is not supported
      content:
        application/problem+json:
          schema:
            allOf:
              - { $ref: '#/components/schemas/ProblemDetails' }
              - type: object
                properties:
                  status: { const: 415 }
                  code: { const: AGENDA_ARTIFACT_UNSUPPORTED }
    AgendaLayoutRejected:
      description: File content does not match the recognized SIMEF HTML layout
      content:
        application/problem+json:
          schema:
            allOf:
              - { $ref: '#/components/schemas/ProblemDetails' }
              - type: object
                properties:
                  status: { const: 422 }
                  code: { const: AGENDA_LAYOUT_REJECTED }
    AgendaImportFailed:
      description: Unexpected technical failure; rollback performed
      content:
        application/problem+json:
          schema:
            allOf:
              - { $ref: '#/components/schemas/ProblemDetails' }
              - type: object
                properties:
                  status: { const: 500 }
                  code: { const: AGENDA_IMPORT_FAILED }
    AgendaImportTimeout:
      description: Import processing timed out; rollback performed
      content:
        application/problem+json:
          schema:
            allOf:
              - { $ref: '#/components/schemas/ProblemDetails' }
              - type: object
                properties:
                  status: { const: 504 }
                  code: { const: AGENDA_IMPORT_TIMEOUT }
    NoActiveAppointments:
      description: No hay citas activas para los servicios solicitados en esta fecha
      content:
        application/problem+json:
          schema:
            allOf:
              - { $ref: '#/components/schemas/ProblemDetails' }
              - type: object
                properties:
                  status: { const: 422 }
                  code: { const: NO_ACTIVE_APPOINTMENTS }